Rename file of component
