+++ src/ggrc/assets/javascripts/components/access_control_list/access-control-list-roles-helper.js
