+++ src/ggrc/assets/javascripts/controllers/modals_controller.js
@@ -3,7 +3,7 @@
  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
 
-import '../components/access_control_list/access_control_list_roles_helper'
+import '../components/access_control_list/access-control-list-roles-helper'
 
 (function (can, $) {
   can.Control('GGRC.Controllers.Modals', {
